requirements.txt
pandas
requests
apache-airflow[postgres]
psycopg2
python-dotenv